pyproject.toml
[tool.poetry]
name = "plants-api"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Liz Dahlstrom"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2" # application framework
sqlmodel = "^0.0.14" # model database tables as classes
pg8000 = "^1.30.5" # postgres connector

[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1" # database migrations
autoflake = "^2.3.1" # remove unused imports
black = "^24.3.0" # linter
factory-boy = "^3.3.0" # fake objects for testing
httpx = "^0.26.0" # async db toolkit
mock = "^5.1.0" # fake responses for tests
mypy = "^1.8.0"   # type checker
pre-commit = "^3.7.0" # run checks before committing
pytest = "^8.0.0" # testing framework
uvicorn = "^0.27.0" # application engine

[tool.autoflake]
check = true
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
ignore-init-module-imports = true
ignore-pass-statements = true
recursive = true
exclude = [".venv", ".vscode", ".pytest_cache", "*.__pycache__", "alembic/*"]
in-place = true

[tool.black]
force-exclude = ["alembic"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
